I'm tired
I’ve been writing code most of the day now. I’m not yet done… I need to
finish all the api controller so that we can hook it up to the UI.
Essentially we can make this work within a week. There’s a lot to do
(thats the only problem)